BackEnd/requirements.txt
fastapi>=0.109.0         # فریم‌ورک جدید شما (جایگزین جنگو)
uvicorn[standard]        # سرور اجرای برنامه (بسیار سریع و Asynchronous)
pydantic-settings        # برای مدیریت تنظیمات و متغیرهای محیطی (.env)
polars                   # پردازش داده (حفظ شود)
connectorx               # خواندن سریع از دیتابیس (حفظ شود - عالی برای FastAPI)
pyarrow                  # موتور پولارز (حفظ شود)
numpy                    # محاسبات ریاضی (حفظ شود)
psycopg2-binary          # درایور دیتابیس (می‌تواند بماند)
adbc-driver-postgresql